Cursor support
Now correctly handles cursor and deleting characters. It seems to work really well.

There is a beforeInput event that you can use to do something before the input is updated, but i’m not sure that is needed as it is working fine now. That said, it could theoretically be optimized using this event trigger.
